rename ChatColumn to MessageColumn
