--- src/Designer/frontend/libs/studio-assistant/src/components/ChatColumn/index.ts
@@ -1,2 +1,0 @@
-export { ChatColumn } from './ChatColumn';
-export type { ChatColumnProps } from './ChatColumn';
