+++ src/Designer/frontend/libs/studio-assistant/src/components/CompleteInterface/CompleteInterface.tsx
@@ -6,7 +6,7 @@ import classes from './CompleteInterface.module.css';
 import { HeadingBar } from '../HeadingBar/HeadingBar';
 import { ThreadColumn } from '../ThreadColumn/ThreadColumn';
 import { ThreadColumnCollapsed } from '../ThreadColumnCollapsed/ThreadColumnCollapsed';
-import { ChatColumn } from '../ChatColumn/ChatColumn';
+import { MessageColumn } from '../MessageColumn/MessageColumn';
 import { ToolColumnMode } from '../../types/ToolColumnMode';
 import type { AssistantProps } from '../../Assistant/Assistant';
 
@@ -85,7 +85,7 @@ export function CompleteInterface({
           )}
         </StudioResizableLayout.Element>
         <StudioResizableLayout.Element minimumSize={400}>
-          <ChatColumn
+          <MessageColumn
             texts={texts}
             messages={messages}
             onSubmitMessage={onSubmitMessage}
