+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/AssistantLoadingBubble/index.ts
