+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/AssistantMessage/AssistantMessage.module.css
