+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/AssistantMessage/FilesChangedList/FilesChangedList.test.tsx
