+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/AssistantMessage/index.ts
