+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/ChatAvatar/ChatAvatar.module.css
