+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/MessageRow/MessageRow.module.css
