+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/MessageRow/MessageRow.test.tsx
