+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/Messages.test.tsx
