+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/Messages/UserMessage/UserMessage.tsx
