+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/UserInput/UserInput.tsx
