+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/UserInput/index.ts
