+++ src/Designer/frontend/libs/studio-assistant/src/components/MessageColumn/index.ts
@@ -1,0 +1,2 @@
+export { MessageColumn } from './MessageColumn';
+export type { MessageColumnProps } from './MessageColumn';
