+++ src/Designer/frontend/libs/studio-assistant/src/index.ts
@@ -8,7 +8,7 @@ export type {
   AboutAssistantDialogTexts,
   MessageFeedbackTexts,
 } from './types/AssistantTexts';
-export type { MessagesProps } from './components/ChatColumn/Messages/Messages';
+export type { MessagesProps } from './components/MessageColumn/Messages/Messages';
 export type { AssistantMessage } from './types/ChatThread';
 export type { User } from './types/User';
 export type { WorkflowStatus } from './types/WorkflowStatus';
